Removed comments
I realized that the loop the iterates over all the values in the hash
then uses the includes? method which probably makes it O(n^2) or O(n log
n), depending on how the includes? method works. For this reason, the
comments discussing the time complexity of this program have been
removed.